Minor cleaning up
  - find_itype now in read.tools
  - some minor adjustments to tests
